# kinect-api-contracts/src/main/java/kinect/api/contracts/kinect-trainingprograms/openapi.yaml
openapi: 3.0.3
info:
  title: Kinect Tech - Training Programs API Contracts
  version: 1.0.0
paths:
  /api/v1/training-programs:
    post:
      summary: Create a training program
      operationId: createTrainingProgram
      tags:
        - TrainingPrograms
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrainingProgramRequest'
            example:
              studentId: 1
              trainerId: 2
              name: Treino A - Corpo inteiro
              exercises:
                - name: Agachamento livre
                  sets: 3
                  repetitions: 12
                  restSeconds: 60
                  notes: Manter a coluna alinhada
                - name: Supino reto
                  sets: 3
                  repetitions: 10
                  restSeconds: 90
      responses:
        '201':
          description: Training program created
    get:
      summary: List training programs
      operationId: getTrainingPrograms
      tags:
        - TrainingPrograms
      responses:
        '200':
          description: Training programs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TrainingProgram'
              example:
                - id: 1
                  studentId: 1
                  trainerId: 2
                  name: Treino A - Corpo inteiro
                  exercises:
                    - name: Agachamento livre
                      sets: 3
                      repetitions: 12
                      restSeconds: 60
                      notes: Manter a coluna alinhada
                    - name: Supino reto
                      sets: 3
                      repetitions: 10
                      restSeconds: 90
                  createdAt: '2026-10-03T12:00:00Z'
                  updatedAt: '2026-10-03T12:00:00Z'
  /api/v1/training-programs/{trainingProgramId}:
    parameters:
      - name: trainingProgramId
        in: path
        required: true
        schema:
          type: integer
          format: int64
        example: 1
    get:
      summary: Get training program by id
      operationId: getTrainingProgramById
      tags:
        - TrainingPrograms
      responses:
        '200':
          description: Training program found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrainingProgram'
              example:
                id: 1
                studentId: 1
                trainerId: 2
                name: Treino A - Corpo inteiro
                exercises:
                  - name: Agachamento livre
                    sets: 3
                    repetitions: 12
                    restSeconds: 60
                    notes: Manter a coluna alinhada
                  - name: Supino reto
                    sets: 3
                    repetitions: 10
                    restSeconds: 90
                createdAt: '2026-10-03T12:00:00Z'
                updatedAt: '2026-10-03T12:00:00Z'
        '404':
          description: Training program not found
    put:
      summary: Update a training program
      operationId: updateTrainingProgram
      tags:
        - TrainingPrograms
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrainingProgramRequest'
            example:
              studentId: 1
              trainerId: 2
              name: Treino A - Corpo inteiro
              exercises:
                - name: Agachamento livre
                  sets: 3
                  repetitions: 12
                  restSeconds: 60
                  notes: Manter a coluna alinhada
                - name: Supino reto
                  sets: 3
                  repetitions: 10
                  restSeconds: 90
      responses:
        '204':
          description: Training program updated
        '404':
          description: Training program not found
    delete:
      summary: Delete a training program
      operationId: deleteTrainingProgram
      tags:
        - TrainingPrograms
      responses:
        '204':
          description: Training program deleted
        '404':
          description: Training program not found
components:
  schemas:
    TrainingProgramRequest:
      type: object
      required:
        - studentId
        - trainerId
        - name
        - exercises
      properties:
        studentId:
          type: integer
          format: int64
          example: 1
        trainerId:
          type: integer
          format: int64
          example: 2
        name:
          type: string
          maxLength: 120
          example: Treino A - Corpo inteiro
        exercises:
          type: array
          items:
            $ref: '#/components/schemas/Exercise'
    Exercise:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 120
          example: Agachamento livre
        sets:
          type: integer
          minimum: 1
          example: 3
        repetitions:
          type: integer
          minimum: 1
          example: 12
        restSeconds:
          type: integer
          minimum: 0
          example: 60
        notes:
          type: string
          maxLength: 500
          example: Manter a coluna alinhada
    TrainingProgram:
      allOf:
        - $ref: '#/components/schemas/TrainingProgramRequest'
        - type: object
          properties:
            id:
              type: integer
              format: int64
              example: 1
            createdAt:
              type: string
              format: date-time
              example: '2026-10-03T12:00:00Z'
            updatedAt:
              type: string
              format: date-time
              example: '2026-10-03T12:00:00Z'